api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
snowflake-connector-python==3.7.1
cryptography==42.0.5
python-dotenv==1.0.1
pydantic-settings==2.1.0
cachetools==5.3.2
